build: update dependency management
Change build system to use vcpkg's manifest mode instead. Local binary
caches are now stored locally.

The architecture of the image is now dependent on the architecture of
the docker host. Native images are available for amd64 and arm64. This
improves build time and runtime performance because there is no need
to run an QEMU layer for an image of a different architecture.

Add docker-compose configurations to standardise the launching of
containers for testing with the RIE. Update README to reflect these
changes.